--- a/lib/goals.js
+++ b/lib/goals.js
@@ -306,7 +306,9 @@ class GoalInvert extends Goal {
   }
 
   heuristic (node) {
-    return -this.goal.heuristic(node)
+    // Keep A* properties sane: heuristic must be non-negative and (ideally) admissible.
+    const h = this.goal.heuristic(node)
+    return h > 0 ? 0 : 0 // conservative (or use Math.max(0, -h) if you still want a gradient)
   }
 
   isEnd (node) {