--- a/lib/interactable.json
+++ b/lib/interactable.json
@@ -132,5 +132,5 @@
   "white_shulker_box",
   "wood_button",
   "yellow_bed",
-  "yelow_shulker_box"
+  "yellow_shulker_box"
 ]