--- a/test/internalTest.js
+++ b/test/internalTest.js
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 
 const mineflayer = require('mineflayer')
-const { goals, pathfinder, Movements } = require('mineflayer-pathfinder')
+const { goals, pathfinder, Movements } = require('../index')
 const { Vec3 } = require('vec3')
 const mc = require('minecraft-protocol')
 const assert = require('assert')