… IMPROVEMENTS.md
… docs/mineflayer_api.md
… lib/movements.js
… lib/passableEntities.json
… lib/runtime-sets.js
… package.json
… scripts/generate-lists.js
… test/lists.test.js