@darinkishore 2025-08-16
lemme close this and do smaller PRs then! 

API isn't changed. 

test is i think b/c node version? the failure is 

```
- Assertion:
    - Expected distanceSquared < 0.01
    - Got Bot: (0.5, 3.7967, 13.3095) vs Simulation: (0.5, 3.4952, 13.4593)
    - Squared distance ≈ 0.113 (mainly Y ≈ 0.30, Z ≈ 0.15)
```